requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
cffi==1.11.5
colorama==0.4.1
coverage==4.5.2
cryptography==2.4.2
dj-database-url==0.5.0
Django==2.1.5
django-autofixture==0.12.1
django-debug-toolbar==1.11
django-extensions==2.1.4
django-nose==1.4.6
djangorestframework==3.9.0
idna==2.8
isort==4.3.4
jwt==0.5.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
pinocchio==0.4.2
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
python-decouple==3.1
pytz==2018.7
six==1.12.0
sqlparse==0.2.4
typed-ast==1.1.1
wrapt==1.10.11
